fix image working

+++ srcs/angles_basic.c
@@ -6,17 +6,17 @@
 /*   By: anbourge <anbourge@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/07 18:08:06 by requinch          #+#    #+#             */
-/*   Updated: 2022/07/09 16:51:04 by anbourge         ###   ########.fr       */
+/*   Updated: 2022/07/20 16:41:04 by anbourge         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/cub3d.h"
 
-float	degtorad(short deg)
+float	degtorad(float deg)
 {
 	float	f;
 	
-	f = deg * (PI / 180);
+	f = deg * (PI / 180.0);
 	return (f);
 }
